+++ rc_lib/style/animation.py
@@ -1,0 +1,5 @@
+"""
+    Style elements intrinsically linked with animation, such as timing.
+"""
+
+END_DELAY = 3
