+++ rc_lib/style/text.py
@@ -1,0 +1,7 @@
+import enum
+
+# Default sizes were 24, 32, 48
+class FontSize(enum.IntEnum):
+    SMALL = 20
+    MEDIUM = 24
+    LARGE = 32
