+++ rc_lib/view_utils/title_sequence.py
@@ -1,7 +1,7 @@
 from manim import *
 from typing import List
 
-from rc_lib import style
+from rc_lib.style import color
 
 __all__ = ["TitleSequence"]
 
@@ -12,7 +12,7 @@ class TitleSequence:
     """
 
     def __init__(
-        self, add_numbers: bool = True, default_color: style.Color = style.DEFAULT_COLOR
+        self, add_numbers: bool = True, default_color: color.Color = color.FOREGROUND
     ) -> None:
         self._add_numbers = add_numbers
         self._default_color = default_color
@@ -24,7 +24,7 @@ class TitleSequence:
         """
         self._number = 1
 
-    def next(self, title: str, color: style.Color | None = None) -> Animation:
+    def next(self, title: str, color: color.Color | None = None) -> Animation:
         text = self._make_text(title, self._default_color if color is None else color)
         self._number += 1
         if self._number == 2:
@@ -33,6 +33,6 @@ class TitleSequence:
         else:
             return Transform(self._first, text, run_time=0.75)
 
-    def _make_text(self, title: str, color: style.Color) -> Text:
+    def _make_text(self, title: str, color: color.Color) -> Text:
         prefix = str(self._number) + ": " if self._add_numbers else ""
         return Text(prefix + title, color=color).to_corner(UP + LEFT)
